Finish writing Cao's method example.
Similar tabular approach to Huang's method. Might actually switch to the
toy example (from k-modes issue) as it is very small and it's far easier
to see where things are coming from. Most of the ground work is done for
these examples now anyway, so may that would be best...