Rewrite results section spiel, finish definitions
After discussion with Vince and Jon, the methods by which our results
are found are now split in two: predictive classification quality and
cluster validity. There are still issues around how to practically
implement accuracy/precision/recall metrics so we are still using the
metrics from before as if they were to be used as classification
metrics. The clustering metrics used are simply the cost function of the
algorithm, and the silhouette of a clustering.