Finish elbow subsec; begin nclasses.
Elbow method section is pretty much done. There is a note on potential
further theoretic work at the end that should be addressed.

Beginning the classes subsection, tables have been summarised but
figures must be addressed. Holding off on this until meeting on
2020-01-21 as figures indicate k-modes terminates on a worse solution
than the original. Again, this is a point that should be addressed as it
is unclear whether it is theoretical or an implementation issue.